+++ nanome_minimization/menu.py
@@ -26,7 +26,7 @@ class MinimizationMenu():
         self.__update_start_btn(running)
 
     def toggle_minimization(self):
-        if self.__plugin._process._is_running:
+        if self.__plugin._process.is_running:
             self.stop_minimization()
         else:
             self.start_minimization()
@@ -81,7 +81,7 @@ class MinimizationMenu():
             self.toggle_minimization()
 
         # loading menus
-        menu = nanome.ui.Menu.io.from_json(os.path.join(os.path.dirname(__file__), "_MinimizationMenu.json"))
+        menu = nanome.ui.Menu.io.from_json(os.path.join(os.path.dirname(__file__), "minimization_menu.json"))
         self.__menu = menu
         self.__plugin.menu = menu
